[dtensor] move some modules to private namespace
as titled, moving some modules that are mainly for DTensor private usage
to be a private module.

[ghstack-poisoned]
