+++ test/distributed/_tensor/test_common_rules.py
@@ -3,7 +3,7 @@
 
 import torch
 from torch.distributed._tensor import DeviceMesh
-from torch.distributed._tensor.op_schema import OpSchema
+from torch.distributed._tensor._op_schema import OpSchema
 
 from torch.distributed._tensor.ops.common_rules import einop_rule, pointwise_rule
 from torch.distributed._tensor.placement_types import DTensorSpec, TensorMeta
