+++ test/distributed/_tensor/test_redistribute.py
@@ -206,7 +206,7 @@ class RedistributeTest(DTensorTestBase):
         with self.assertRaisesRegex(RuntimeError, "Can not redistribute to _Partial"):
             partial_tensor = replica_tensor.redistribute(device_mesh, [partial_spec])
 
-        from torch.distributed._tensor.redistribute import Redistribute
+        from torch.distributed._tensor._redistribute import Redistribute
 
         comm_mode = CommDebugMode()
