+++ torch/distributed/_spmd/distribute.py
@@ -14,7 +14,8 @@ from torch.distributed._spmd.graph_utils import OP
 from torch.distributed._spmd.log_utils import get_logger
 
 from torch.distributed._tensor import DeviceMesh, DTensor
-from torch.distributed._tensor.op_schema import OpSchema
+from torch.distributed._tensor._op_schema import OpSchema
+from torch.distributed._tensor._redistribute import redistribute_local_tensor
 from torch.distributed._tensor.placement_types import (
     _Partial,
     DTensorSpec,
@@ -23,7 +24,6 @@ from torch.distributed._tensor.placement_types import (
     Shard,
     TensorMeta,
 )
-from torch.distributed._tensor.redistribute import redistribute_local_tensor
 from torch.fx.experimental.proxy_tensor import make_fx, proxy_slot
 from torch.utils import _pytree as pytree
 from torch.utils._pytree import tree_flatten, tree_map, tree_map_only, tree_unflatten
