+++ torch/distributed/_tensor/_dispatch.py
@@ -10,22 +10,22 @@ import torch
 import torch.distributed as dist
 import torch.distributed._tensor.api as dtensor
 import torch.distributed._tensor.random as random
-from torch.distributed._tensor._utils import try_find_mesh_from_args
-from torch.distributed._tensor.op_schema import (
+from torch.distributed._tensor._op_schema import (
     _is_inplace_op,
     _is_out_variant_op,
     OpInfo,
     OpSchema,
     OutputSpecType,
 )
-from torch.distributed._tensor.placement_types import DTensorSpec, Replicate, TensorMeta
-from torch.distributed._tensor.random import is_rng_supported_mesh
-from torch.distributed._tensor.redistribute import redistribute_local_tensor
-from torch.distributed._tensor.sharding_prop import ShardingPropagator
-from torch.distributed._tensor.tp_conv import (
+from torch.distributed._tensor._redistribute import redistribute_local_tensor
+from torch.distributed._tensor._sharding_prop import ShardingPropagator
+from torch.distributed._tensor._tp_conv import (
     convolution_backward_handler,
     convolution_handler,
 )
+from torch.distributed._tensor._utils import try_find_mesh_from_args
+from torch.distributed._tensor.placement_types import DTensorSpec, Replicate, TensorMeta
+from torch.distributed._tensor.random import is_rng_supported_mesh
 
 if TYPE_CHECKING:
     from torch.distributed.device_mesh import DeviceMesh
