+++ torch/distributed/_tensor/_sharding_prop.py
@@ -5,12 +5,7 @@ from typing import Callable, cast, Dict, List, Optional, Sequence, Tuple, Union
 import torch
 from torch._ops import OpOverload
 from torch._subclasses import FakeTensorMode
-from torch.distributed._tensor._utils import (
-    compute_local_shape,
-    compute_local_stride,
-    try_find_mesh_from_args,
-)
-from torch.distributed._tensor.op_schema import (
+from torch.distributed._tensor._op_schema import (
     OpInfo,
     OpSchema,
     OpStrategy,
@@ -20,6 +15,11 @@ from torch.distributed._tensor.op_schema import (
     RuntimeSchemaInfo,
     StrategyType,
     TupleStrategy,
+)
+from torch.distributed._tensor._utils import (
+    compute_local_shape,
+    compute_local_stride,
+    try_find_mesh_from_args,
 )
 from torch.distributed._tensor.placement_types import DTensorSpec, TensorMeta
 from torch.distributed.device_mesh import DeviceMesh
