+++ torch/distributed/_tensor/ops/basic_strategy.py
@@ -3,7 +3,7 @@ from dataclasses import dataclass
 
 from typing import List, Set, Tuple
 
-from torch.distributed._tensor.op_schema import OpStrategy, PlacementStrategy
+from torch.distributed._tensor._op_schema import OpStrategy, PlacementStrategy
 from torch.distributed._tensor.placement_types import (
     _Partial,
     DTensorSpec,
