+++ torch/distributed/_tensor/ops/common_rules.py
@@ -2,13 +2,13 @@
 from typing import cast, Dict, List, Optional, Tuple
 
 import torch
-from torch.distributed._tensor._utils import compute_local_shape
-from torch.distributed._tensor.op_schema import (
+from torch.distributed._tensor._op_schema import (
     _is_inplace_op,
     _is_out_variant_op,
     OpSchema,
     OutputSharding,
 )
+from torch.distributed._tensor._utils import compute_local_shape
 from torch.distributed._tensor.ops.utils import prod
 from torch.distributed._tensor.placement_types import DTensorSpec, TensorMeta
