+++ torch/distributed/_tensor/ops/experimental_ops.py
@@ -8,7 +8,7 @@ except ModuleNotFoundError:
     np = None  # type: ignore[assignment]
 
 import torch
-from torch.distributed._tensor.op_schema import OpSchema, OutputSharding
+from torch.distributed._tensor._op_schema import OpSchema, OutputSharding
 from torch.distributed._tensor.ops.utils import register_prop_rule
 from torch.distributed._tensor.placement_types import DTensorSpec, TensorMeta
