+++ torch/distributed/_tensor/ops/math_ops.py
@@ -7,7 +7,7 @@ from typing import cast, List, Optional, Sequence, Tuple, Union
 
 import torch
 
-from torch.distributed._tensor.op_schema import (
+from torch.distributed._tensor._op_schema import (
     OpSchema,
     OpStrategy,
     PlacementStrategy,
