+++ torch/distributed/_tensor/ops/tensor_ops.py
@@ -4,7 +4,7 @@ from typing import cast, List, Optional, Sequence, Tuple
 
 import torch
 
-from torch.distributed._tensor.op_schema import (
+from torch.distributed._tensor._op_schema import (
     OpSchema,
     OpStrategy,
     OutputSharding,
