+++ torch/distributed/_tensor/ops/utils.py
@@ -5,8 +5,8 @@ from typing import cast, Iterable, List, Sequence, Tuple, Union
 
 import torch
 from torch.distributed._tensor._collective_utils import redistribute_cost
+from torch.distributed._tensor._op_schema import OpStrategy, RuntimeSchemaInfo
 from torch.distributed._tensor.api import DTensor
-from torch.distributed._tensor.op_schema import OpStrategy, RuntimeSchemaInfo
 from torch.distributed._tensor.placement_types import (
     _Partial,
     DTensorSpec,
