+++ torch/distributed/_tensor/random.py
@@ -11,7 +11,7 @@ from torch.distributed._tensor.placement_types import DTensorSpec, Shard
 from torch.distributed.device_mesh import _get_device_handle, DeviceMesh
 
 
-_rng_tracker: Optional["RNGStateTracker"] = None
+_rng_tracker: Optional["_RNGStateTracker"] = None
 
 
 def is_rng_supported_mesh(device_mesh: DeviceMesh) -> bool:
@@ -91,9 +91,9 @@ def manual_seed(seed: int, device_mesh: DeviceMesh) -> None:
             )
 
 
-class RNGStateTracker:
+class _RNGStateTracker:
     """
-    RNGStateTracker stores Random Number Generator (RNG) state (a ByteTensor object)
+    _RNGStateTracker stores Random Number Generator (RNG) state (a ByteTensor object)
     in a dict, mapping from a corresponding tag to each state tensor. It also provides
     a set of convenient utility methods to help access/modify the state tensors. The most
     important interface is _distribute_region which will be used when DTensor executes
@@ -145,9 +145,9 @@ class RNGStateTracker:
         pass
 
 
-class OffsetBasedRNGTracker(RNGStateTracker):
+class OffsetBasedRNGTracker(_RNGStateTracker):
     """
-    This subclass of `RNGStateTracker` defines the default policy of how RNG states
+    This subclass of `_RNGStateTracker` defines the default policy of how RNG states
     should be shared and synchronized among all ranks to respect the semantics of DTensor
     random operators.
     """
@@ -331,7 +331,7 @@ class OffsetBasedRNGTracker(RNGStateTracker):
         return shard_linear_idx
 
 
-class TensorParallelRNGTracker(RNGStateTracker):
+class TensorParallelRNGTracker(_RNGStateTracker):
     def __init__(self, device_type: str = "cuda"):
         super().__init__(device_type)
         # copy the default RNG state
